rewrite rest components/showobjects
